+++ solr/core.py
@@ -21,8 +21,9 @@ from .utils import (
 )
 from .response import Response, Results
 from .parsers import parse_json_response, parse_query_response
+from .schema import SchemaAPI
 
-__version__ = "1.3.0"
+__version__ = "1.4.0"
 
 __all__ = ['SolrException', 'SolrVersionError', 'Solr',
            'Response', 'SearchHandler']
@@ -121,6 +122,7 @@ class Solr:
         self.debug = debug
         self.select = SearchHandler(self, "/select")
         self.mlt = SearchHandler(self, "/mlt")
+        self.schema = SchemaAPI(self)
         self.server_version = self._detect_version()
 
     def close(self) -> None: